first version of changing the data type

+++ src/cajal/EMD.h
@@ -30,7 +30,7 @@ enum ProblemType {
 };
 
 int myCMessage(int a,int b);
-int EMD_wrap(int n1,int n2, double *X, double *Y,double *D, double *G, double* alpha, double* beta, double *cost, uint64_t maxIter);
+int EMD_wrap(int n1,int n2, float *X, float *Y,float *D, float *G, float* alpha, float* beta, float *cost, uint64_t maxIter);
 int EMD_wrap_omp(int n1,int n2, double *X, double *Y,double *D, double *G, double* alpha, double* beta, double *cost, uint64_t maxIter, int numThreads);
